Cleanup warnings
git-svn-id: https://svn.java.net/svn/electric~svn/trunk@6091 b64ed6c1-75a8-e629-b1da-d0dae74a743a

+++ srcj/com/sun/electric/tool/io/output/ReadableDump.java
@@ -118,7 +118,7 @@ public class ReadableDump extends ELIB {
             CellBackup cellBackup = snapshot.getCell(cellId);
             startCell(cellBackup, 0);
             int groupIndex = groupRenumber[snapshot.cellGroups[cellId.cellIndex]];
-            printWriter.println("***cell: " + ((Integer)entry.getValue()).intValue() + "/" + groupIndex);
+            printWriter.println("***cell: " + entry.getValue().intValue() + "/" + groupIndex);
             writeCellInfo(cellBackup);
             
             // write tool information
